+++ src/Debug/TraceableHub.php
@@ -16,6 +16,7 @@ namespace Symfony\Component\Mercure\Debug;
 use Symfony\Component\Mercure\HubInterface;
 use Symfony\Component\Mercure\Jwt\TokenFactoryInterface;
 use Symfony\Component\Mercure\Jwt\TokenProviderInterface;
+use Symfony\Component\Mercure\ProtocolVersion;
 use Symfony\Component\Mercure\RemoteHubInterface;
 use Symfony\Component\Mercure\Update;
 use Symfony\Component\Stopwatch\Stopwatch;
@@ -64,6 +65,16 @@ final class TraceableHub implements RemoteHubInterface, ResetInterface
     public function getFactory(): ?TokenFactoryInterface
     {
         return $this->hub->getFactory();
+    }
+
+    public function getProtocolVersion(): ProtocolVersion
+    {
+        return $this->hub->getProtocolVersion();
+    }
+
+    public function getCookieName(): string
+    {
+        return $this->hub->getCookieName();
     }
 
     public function publish(Update $update): string