+++ tests/Jwt/FactoryTokenProviderTest.php
@@ -17,6 +17,7 @@ use Lcobucci\JWT\Signer\Key;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Mercure\Jwt\FactoryTokenProvider;
 use Symfony\Component\Mercure\Jwt\LcobucciFactory;
+use Symfony\Component\Mercure\Jwt\TokenFactoryInterface;
 
 final class FactoryTokenProviderTest extends TestCase
 {
@@ -33,5 +34,18 @@ final class FactoryTokenProviderTest extends TestCase
             'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJtZXJjdXJlIjp7InB1Ymxpc2giOlsiKiJdLCJzdWJzY3JpYmUiOltdfX0.ZTK3JhEKO1338LAgRMw6j0lkGRMoaZtU4EtGiAylAns',
             $provider->getJwt()
         );
+    }
+
+    public function testAdditionalClaimsAreForwardedToTheFactory()
+    {
+        $factory = $this->createMock(TokenFactoryInterface::class);
+        $factory
+            ->expects($this->once())
+            ->method('create')
+            ->with(['a'], ['b'], ['iss' => 'https://example.com'])
+        ;
+
+        $provider = new FactoryTokenProvider($factory, ['a'], ['b'], ['iss' => 'https://example.com']);
+        $provider->getJwt();
     }
 }